fix: move feature flag component into a separate file
